백오피스 리뷰 댓글 생성 수정

+++ domain-review/src/main/java/co/dalicious/domain/review/dto/ReviewAdminResDto.java
@@ -15,7 +15,7 @@ import java.util.stream.Collectors;
 @Builder
 public class ReviewAdminResDto {
 
-    private Integer unansweredCount;
+    private Long unansweredCount;
     private List<MakersInfo> makersInfoList;
     private List<ReviewList> reviewList;
 
@@ -70,7 +70,7 @@ public class ReviewAdminResDto {
         }
     }
 
-    public static ReviewAdminResDto create(List<Makers> makersList, List<ReviewList> reviewList, Integer count) {
+    public static ReviewAdminResDto create(List<Makers> makersList, List<ReviewList> reviewList, Long count) {
         List<MakersInfo> makersInfos = makersList.stream().map(MakersInfo::create).collect(Collectors.toList());
         return ReviewAdminResDto.builder()
                 .unansweredCount(count)
